remove ivKeyGenerator and encryption method
